requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Deep learning and transformers
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
protobuf>=3.20.0

# Dimensionality reduction and visualization
scikit-learn>=1.3.0
umap-learn>=0.5.3
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0
psutil>=5.9.0  # Memory monitoring

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing (optional)
pytest>=7.4.0
